frontend testing refactoring
